refactor: extract command handlers to commands module, improve CLI type safety, and add security validations
