+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod cli;
+pub mod commands;
 pub mod git;
 pub mod installer;
 pub mod manifest;
